commit before pulling master
